Make feature flags simple again
