+++ apps/web/.well-known/vercel/flags/route.ts
@@ -1,0 +1,3 @@
+import { getFlags } from '@repo/feature-flags/access';
+
+export const GET = getFlags;
