+++ packages/feature-flags/index.ts
@@ -1,33 +1,3 @@
-import { auth } from '@clerk/nextjs/server';
-import { analytics } from '@repo/design-system/lib/analytics/server';
-import { type FlagOverridesType, decrypt } from '@vercel/flags';
-import { unstable_flag as flag } from '@vercel/flags/next';
+import { createFlag } from './create-flag';
 
-export const showBetaFeature = flag({
-  key: 'showBetaFeature',
-  async decide(params) {
-    const { key } = this;
-    const overrideCookie = params.cookies.get('vercel-flag-overrides')?.value;
-    const overrides = overrideCookie
-      ? await decrypt<FlagOverridesType>(overrideCookie)
-      : {};
-
-    if (overrides && key in overrides) {
-      return overrides[key];
-    }
-
-    try {
-      const { userId } = await auth();
-
-      if (!userId) {
-        return false;
-      }
-
-      const isEnabled = await analytics.isFeatureEnabled(key, userId);
-
-      return isEnabled ?? false;
-    } catch (_error) {
-      return false;
-    }
-  },
-});
+export const showBetaFeature = createFlag('showBetaFeature');
